+++ src/main/java/com/atlauncher/gui/card/packbrowser/ATLauncherPackCard.java
@@ -56,12 +56,10 @@ public class ATLauncherPackCard extends JPanel implements RelocalizationListener
     private final JButton serversButton = new JButton(GetText.tr("Servers"));
     private final JButton modsButton = new JButton(GetText.tr("View Mods"));
     private final Pack pack;
-    private final boolean featured;
 
-    public ATLauncherPackCard(final Pack pack, final boolean featured) {
+    public ATLauncherPackCard(final Pack pack) {
         super();
         this.pack = pack;
-        this.featured = featured;
         setLayout(new BorderLayout());
         setBorder(BorderFactory.createTitledBorder(null, pack.name, TitledBorder.LEADING, TitledBorder.DEFAULT_POSITION,
                 App.THEME.getBoldFont().deriveFont(15f)));
@@ -146,7 +144,7 @@ public class ATLauncherPackCard extends JPanel implements RelocalizationListener
                         .setContent(GetText.tr("Cannot create instance as you have no account selected."))
                         .setType(DialogManager.ERROR).show();
             } else {
-                Analytics.sendEvent(pack.getName(), "Install", featured ? "ATLauncherFeaturedPack" : "ATLauncherPack");
+                Analytics.sendEvent(pack.getName(), "Install", "ATLauncherPack");
                 new InstanceInstallerDialog(pack);
             }
         });
@@ -170,8 +168,7 @@ public class ATLauncherPackCard extends JPanel implements RelocalizationListener
                         .setContent(GetText.tr("Cannot create server as you have no account selected."))
                         .setType(DialogManager.ERROR).show();
             } else {
-                Analytics.sendEvent(pack.getName(), "ServerInstall",
-                        featured ? "ATLauncherFeaturedPack" : "ATLauncherPack");
+                Analytics.sendEvent(pack.getName(), "ServerInstall", "ATLauncherPack");
                 new InstanceInstallerDialog(pack, true);
             }
         });
@@ -187,7 +184,7 @@ public class ATLauncherPackCard extends JPanel implements RelocalizationListener
                         Constants.SERVERS_LIST_PACK, pack.getSafeName())));
 
         this.modsButton.addActionListener(e -> {
-            Analytics.sendEvent(pack.getName(), "ViewMods", featured ? "ATLauncherFeaturedPack" : "ATLauncherPack");
+            Analytics.sendEvent(pack.getName(), "ViewMods", "ATLauncherPack");
             new ViewModsDialog(pack).setVisible(true);
         });
     }